feat: Add default popup to browser action
This commit updates the `manifest.json` to specify a default popup for the browser action.

By setting the `default_popup` property to `menu/index.html`, the extension will now display this HTML file when the user clicks the extension's icon in the toolbar. This establishes the main user interface for the extension.